refactor: extract rich-text-editor logic into reusable mixin
