+++ packages/rich-text-editor/package.json
@@ -39,6 +39,7 @@
     "polymer"
   ],
   "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
     "@polymer/polymer": "^3.0.0",
     "@vaadin/button": "24.3.0-alpha1",
     "@vaadin/component-base": "24.3.0-alpha1",
